third_party/blink/web_tests/external/wpt/service-workers/service-worker/about-blank-replacement.https-expected.txt

This is a testharness.js-based test.
[FAIL] Initial about:blank is controlled, exposed to clients.matchAll(), and matches final Client.
  assert_false: result: failure: could not find about:blank client expected false got true
[FAIL] Initial about:blank modified by parent is controlled, exposed to clients.matchAll(), and matches final Client.
  assert_false: result: failure: could not find about:blank client expected false got true
[FAIL] Popup initial about:blank is controlled, exposed to clients.matchAll(), and matches final Client.
  assert_false: result: failure: could not find about:blank client expected false got true
[FAIL] Simple about:blank is controlled and is exposed to clients.matchAll().
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'scriptURL')"
[FAIL] Nested about:srcdoc is controlled and is exposed to clients.matchAll().
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'scriptURL')"
[FAIL] Dynamic about:blank is controlled and is exposed to clients.matchAll().
  promise_test: Unhandled rejection with value: object "TypeError: Cannot read properties of null (reading 'scriptURL')"
Harness: the test ran to completion.
